updated regular expression for lastkill
There was a bug in the expression where it captured exp as level, instead of the actual level # for exp. Updated the expression to carefully capture the monster name (up to 30 characters), level, total exp, base exp, and area it's from. Blank space is no longer captured, level is now second captured group. Will update cp_mobTableFiller.lua to reflect this change.